requirements.txt
tensorflow==2.20.0
numpy>=1.25.0
pandas==2.3.2
matplotlib==3.10.5
seaborn>=0.12.2
scikit-learn>=1.3.1
Pillow>=10.0.0
ipywidgets>=8.2.0
h5py>=3.9.0
protobuf>=4.24.0
packaging>=23.3
kaggle==1.7.4.5
